Move TestBuildChainFunction to its own class
